Create README.md
This is me on playing "White hacking" just for learning purpose, no offending code will be sent to anyone.
All the code is TDD crafted as usual and tested on Kali Linux image on Virtualbox.
Do not use this code for making any offence to anyone, this is only for learning and understand the intricacy of system and net communication programming, easily developed using a language as Go. I think Go is really one of the best language for doing this.